Refactor payment entity identifiers to use BigInteger for improved precision and consistency

+++ payment-engine/src/main/java/io/gomobi/payment/service/PaymentInitialStatePersister.java
@@ -70,7 +70,7 @@ public class PaymentInitialStatePersister {
                 .merchant(merchant)
                 .subMerchant(subMerchant)
                 .paymentMethod(resolvedHost.paymentMethod())
-                .transactionAmount(BigInteger.valueOf(request.getAmount()))
+                .transactionAmount(request.getAmount())
                 .currencyCode(request.getCurrency())
                 .build();
